Add doc comments to math library internals

+++ source/pervasive/nonlinear_arith/internals/mod_internals_nonlinear.rs
@@ -1,35 +1,57 @@
+//! This file contains proofs related to modulo that require
+//! nonlinear-arithmetic reasoning to prove. These are internal
+//! functions used within the math standard library.
+//!
+//! It's based on the following file from the Dafny math standard library:
+//! `Source/DafnyStandardLibraries/src/Std/Arithmetic/Internal/ModInternalsNonlinear.dfy`.
+//! That file has the following copyright notice:
+//! /*******************************************************************************
+//! *  Original: Copyright (c) Microsoft Corporation
+//! *  SPDX-License-Identifier: MIT
+//! *  
+//! *  Modifications and Extensions: Copyright by the contributors to the Dafny Project
+//! *  SPDX-License-Identifier: MIT 
+//! *******************************************************************************/
+
 #[allow(unused_imports)]
 use builtin::*;
 use builtin_macros::*;
 
 verus! {
 
+/// Computes `x % y`. This is useful where we want to trigger on a
+/// modulo operator but we need a functional rather than a
+/// mathematical trigger. (A trigger must be fully functional or fully
+/// mathematical.)
 pub open spec fn modulus (x: int, y: int) -> int
 {
     x % y
 }
 
-/* the remainder 0 divided by an integer is 0 */
-// #[verifier::spinoff_prover]
+/// Proof that 0 modulo any positive integer `m` is 0
 proof fn lemma_mod_of_zero_is_zero(m:int)
-    requires (0 as int) < m
-    ensures   0 as int % m == 0 as int
+    requires
+        0 < m,
+    ensures
+        0 as int % m == 0 as int,
 {}
 
-/* describes fundamentals of the modulus operator */
+/// Proof of the fundamental theorem of division and modulo: That for
+/// any positive divisor `d` and any integer `x`, `x` is equal to `d`
+/// times `x / d` plus `x % d`.
 #[verifier(nonlinear)]
 pub proof fn lemma_fundamental_div_mod(x:int, d:int)
     requires d != 0
     ensures  x == d * (x / d) + (x % d)
 {}
 
-/* the remained of 0 divided by any integer is always 0 */
-// #[verifier::spinoff_prover]
+/// Proof that 0 modulo any integer is 0
 proof fn lemma_0_mod_anything()
     ensures forall |m: int| m > 0 ==> #[trigger] modulus(0, m) == 0
 {}
 
-/* a natural number x divided by a larger natural number gives a remainder equal to x */
+/// Proof that a natural number `x` divided by a larger natural number
+/// `m` gives a remainder equal to `x`
 #[verifier(nonlinear)]
 pub proof fn lemma_small_mod(x:nat, m:nat)
     requires 
@@ -39,8 +61,8 @@ pub proof fn lemma_small_mod(x:nat, m:nat)
         #[trigger] modulus(x as int, m as int) == x as int
 {}
 
-/* the range of the modulus of any integer will be [0, m) where m is the divisor */
-// Euclid's division lemma
+/// Proof of Euclid's division lemma, i.e., that any integer `x`
+/// modulo any positive integer `m` is in the half-open range `[0, m)`.
 #[verifier(nonlinear)]
 pub proof fn lemma_mod_range(x:int, m:int)
     requires m > 0
